common: Keep fuse-overlayfs for compatibility and rootless
podman now defaults to native overlayfs for the storage driver and
removed the dependency on fuse-overlayfs in containers-common.

We had manually kept it in Fedora 40 as it was released with
fuse-overlayfs so we had to keep it for compatibility during the major
release lifetime.

From discussions, it appears that this driver is still useful in
rootless container use cases and has better performance there, thus
let's keep it in the Atomic Desktops for now as this is a common use
case.

See: https://src.fedoraproject.org/rpms/containers-common/c/447945e59a01cb6715ed2a21877d45bf0b91ef67
See: https://github.com/coreos/fedora-coreos-tracker/issues/1749