common: Add selected packages from anaconda-tools comps group
Import packages from this group currently missing in Atomic Desktops. In
the future, we may rely on the comps group once we have cleaned it up.

Drop bcache-tools from kinoite-common. It is no longer (not sure if it
ever was) part of the anaconda-tools group.

See: https://gitlab.com/fedora/ostree/sig/-/issues/5