kinoite-common: Set update frequency to weekly for auto updates
We had previously enabled auto updates in Discover but we did not set an
update frequency. Kinoite is built daily but it is not necessary for
users to update that often. Set the update frequency to weekly to keep a
reasonable amount of time between updates while making sure that they
happen eventually.

Users may also manually update at anytime if they want updates to happen
faster or change the default in the system settings.

Flatpaks could (and maybe should) be updated faster (daily or every two
days) but we need support for that in Discover [1].

[1] https://bugs.kde.org/show_bug.cgi?id=454422

See: https://fedoraproject.org/wiki/Changes/KDEKinoiteAutoUpdateByDefault