bootupd: Force enable BLS support in GRUB
Until Fedora 41, the bootloader entries were duplicated in the GRUB boot
menu as both the BLS configs and the ostree-grub2 static entries where
included and loaded in the GRUB config.

With the bootloader being updated and system being migrated to a static
GRUB config as part of the composefs support, we now want to rely on BLS
configs only.

The Fedora 31 release notes [1] suggested disabling BLS support in GRUB
to remove the duplicate entries.

We thus need to make sure that this support is turned back in the GRUB
config before systems are updated to F42.

[1]: https://fedoraproject.org/wiki/Common_F31_bugs#On_Fedora_Silverblue.2FIoT.2C_the_GRUB_menu_shows_duplicate_entries

See: https://gitlab.com/fedora/ostree/sig/-/issues/35
See: https://github.com/coreos/bootupd/issues/857