fedora: Enable fedora-third-party-refresh unit
Enable the fedora-third-party-refresh.service unit to make sure that
Fedora Third Party repos configuration updates are applied after
updates, on boot.

On package mode systems, updates to the Fedora Third Party configuration
and repos are triggered on package updates & changes using scriplets and
tansfiletriggerin triggers.

None of those will apply to Atomic Desktops thus this service unit was
created, but never enabled.

We need this enabled to make sure that the Google Chrome repo (disabled
in the following commit) is disabled on rebase to Fedora 43, before any
package layering operation is triggered.