common: Remove plocate
Remove `plocate` as this is a command line tool and it comes with a
service that periodically scans the entire filesystem. As we use
containers and Flatpaks, this list of files is not really useful nor
accurate and changes frequently.

We expect users to use graphical tools instead to help them find their
files if needed.

See: https://gitlab.com/fedora/ostree/sig/-/issues/81