silverblue: Add openssl
While it is not strictly needed for the default installation, it is a
dependency of a frequently installed GNOME Shell extension: GSConnect.

Let's include it as layering is not a great option, containers won't
work here and sysexts are not ready for this yet.

See: https://pagure.io/fedora-workstation/issue/426
See: https://github.com/fedora-silverblue/issue-tracker/issues/201
See: https://github.com/GSConnect/gnome-shell-extension-gsconnect/wiki/Help#openssl-not-found